image-inspector/image-inspector/requirements.txt
gradio
Pillow

